FIX: #206 투표 결과, 댓글 모두 불러오기 전까지 Loading 화면 띄우기
